backend/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
click==8.0.1
fastapi==0.65.1
h11==0.12.0
pycodestyle==2.7.0
pydantic==1.8.2
PyJWT==2.1.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
